# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tako Master Server Control Plane API
  version: 1.0.0
  description: |
    REST API specification for Tako Master Server.
    Provides cluster orchestration, node management, application lifecycles, and deployment pipelines.
servers:
  - url: http://localhost:8080
    description: Local Master Server
  - url: https://api.gettako.dev
    description: Production Master Server
paths:
...
  /api/v1/auth/login:
    post:
      summary: Login user
      operationId: loginUser
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful with JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
...
  /api/v1/auth/me:
    get:
      summary: Get current authenticated user
      operationId: getCurrentUser
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /api/v1/projects:
    get:
      summary: List projects
      operationId: listProjects
      tags: [Projects]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      operationId: createProject
      tags: [Projects]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /api/v1/projects/{id}:
    get:
      summary: Get project by ID or slug
      operationId: getProject
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    delete:
      summary: Delete project
      operationId: deleteProject
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /api/v1/deployments/{id}/logs:
    get:
      summary: Stream live deployment build logs (SSE)
      operationId: streamDeploymentLogs
      tags: [Deployments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Realtime log chunks stream
          content:
            text/event-stream:
              schema:
                type: string
...
  /api/v1/audit-logs:
    get:
      summary: List audit logs
      operationId: listAuditLogs
      tags: [Audit]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
...
  /api/v1/settings/{key}:
    get:
      summary: Get cluster setting by key
      operationId: getClusterSettingByKey
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting value
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Setting not found
    put:
      summary: Update or create cluster setting by key
      operationId: updateClusterSettingByKey
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: object
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          example: admin
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        environment:
          type: string
        status:
          type: string
...
    AuditActor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
    AuditLog:
      type: object
      properties:
        id:
          type: string
        actor:
          $ref: '#/components/schemas/AuditActor'
        action:
          type: string
          example: create_service
        targetType:
          type: string
          enum: [service, project, node, user, settings]
        targetId:
          type: string
        targetName:
          type: string
        metadata:
          type: object
        ipAddress:
          type: string
        timestamp:
          type: string
...