# server/internal/store/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql"
	"time"
)

type AuditLog struct {
	ID         string    `json:"id"`
	ActorID    string    `json:"actor_id"`
	ActorName  string    `json:"actor_name"`
	ActorEmail string    `json:"actor_email"`
	Action     string    `json:"action"`
	TargetType string    `json:"target_type"`
	TargetID   string    `json:"target_id"`
	TargetName string    `json:"target_name"`
	Metadata   string    `json:"metadata"`
	IpAddress  string    `json:"ip_address"`
	Timestamp  time.Time `json:"timestamp"`
}

type ClusterSetting struct {
	Key       string    `json:"key"`
	Value     string    `json:"value"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Deployment struct {
	ID            string       `json:"id"`
	ServiceID     string       `json:"service_id"`
	CommitHash    string       `json:"commit_hash"`
	CommitMessage string       `json:"commit_message"`
	Branch        string       `json:"branch"`
	Author        string       `json:"author"`
	Status        string       `json:"status"`
	Steps         string       `json:"steps"`
	Logs          string       `json:"logs"`
	DurationMs    int64        `json:"duration_ms"`
	StartedAt     time.Time    `json:"started_at"`
	FinishedAt    sql.NullTime `json:"finished_at"`
	CreatedAt     time.Time    `json:"created_at"`
}

type Node struct {
	ID            string       `json:"id"`
	Name          string       `json:"name"`
	IpAddress     string       `json:"ip_address"`
	PublicIp      string       `json:"public_ip"`
	Role          string       `json:"role"`
	Status        string       `json:"status"`
	CpuTotalCores int64        `json:"cpu_total_cores"`
	MemoryTotalMb int64        `json:"memory_total_mb"`
	DiskTotalGb   float64      `json:"disk_total_gb"`
	CpuPercent    float64      `json:"cpu_percent"`
	MemoryUsedMb  int64        `json:"memory_used_mb"`
	DiskUsedGb    float64      `json:"disk_used_gb"`
	NetworkRxKbps float64      `json:"network_rx_kbps"`
	NetworkTxKbps float64      `json:"network_tx_kbps"`
	DockerVersion string       `json:"docker_version"`
	Os            string       `json:"os"`
	KernelVersion string       `json:"kernel_version"`
	Uptime        string       `json:"uptime"`
	LastHeartbeat sql.NullTime `json:"last_heartbeat"`
	EnrollToken   string       `json:"enroll_token"`
	CreatedAt     time.Time    `json:"created_at"`
	UpdatedAt     time.Time    `json:"updated_at"`
}

type NodeMetric struct {
	ID            int64     `json:"id"`
	NodeID        string    `json:"node_id"`
	CpuPercent    float64   `json:"cpu_percent"`
	MemoryUsedMb  int64     `json:"memory_used_mb"`
	MemoryTotalMb int64     `json:"memory_total_mb"`
	DiskUsedGb    int64     `json:"disk_used_gb"`
	DiskTotalGb   int64     `json:"disk_total_gb"`
	NetworkRxKbps float64   `json:"network_rx_kbps"`
	NetworkTxKbps float64   `json:"network_tx_kbps"`
	RecordedAt    time.Time `json:"recorded_at"`
}

type Project struct {
	ID          string    `json:"id"`
	Name        string    `json:"name"`
	Slug        string    `json:"slug"`
	Description string    `json:"description"`
	Environment string    `json:"environment"`
	Status      string    `json:"status"`
	Tags        string    `json:"tags"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type Service struct {
	ID               string    `json:"id"`
	ProjectID        string    `json:"project_id"`
	NodeID           string    `json:"node_id"`
	Name             string    `json:"name"`
	Slug             string    `json:"slug"`
	Type             string    `json:"type"`
	Status           string    `json:"status"`
	Repository       string    `json:"repository"`
	Branch           string    `json:"branch"`
	CommitHash       string    `json:"commit_hash"`
	Dockerfile       string    `json:"dockerfile"`
	BuildCommand     string    `json:"build_command"`
	ComposeFile      string    `json:"compose_file"`
	Image            string    `json:"image"`
	DatabaseType     string    `json:"database_type"`
	DatabaseVersion  string    `json:"database_version"`
	ConnectionString string    `json:"connection_string"`
	Ports            string    `json:"ports"`
	Replicas         int64     `json:"replicas"`
	CpuLimit         float64   `json:"cpu_limit"`
	MemoryLimitMb    int64     `json:"memory_limit_mb"`
	CreatedAt        time.Time `json:"created_at"`
	UpdatedAt        time.Time `json:"updated_at"`
}

type ServiceDomain struct {
	ID              string    `json:"id"`
	ServiceID       string    `json:"service_id"`
	Domain          string    `json:"domain"`
	Ssl             int64     `json:"ssl"`
	IsPrimary       int64     `json:"is_primary"`
	Port            int64     `json:"port"`
	Path            string    `json:"path"`
	CertificateType string    `json:"certificate_type"`
	CreatedAt       time.Time `json:"created_at"`
}

type ServiceEnvVar struct {
	ID        string    `json:"id"`
	ServiceID string    `json:"service_id"`
	Key       string    `json:"key"`
	Value     string    `json:"value"`
	IsSecret  int64     `json:"is_secret"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type User struct {
	ID           string         `json:"id"`
	Name         string         `json:"name"`
	Email        string         `json:"email"`
	PasswordHash string         `json:"password_hash"`
	Role         string         `json:"role"`
	AvatarUrl    sql.NullString `json:"avatar_url"`
	CreatedAt    time.Time      `json:"created_at"`
	UpdatedAt    time.Time      `json:"updated_at"`
}